Add MultiTrackBar

+++ Real-Time Corruptor/BizHawk_RTC/BizHawk.Client.EmuHawk/RTC/Hooks.cs
@@ -102,7 +102,7 @@ namespace Vanguard
 			CPU_STEP_Count++;
 
 			bool autoCorrupt = CorruptCore.AutoCorrupt;
-			int errorDelay = CorruptCore.ErrorDelay;
+			long errorDelay = CorruptCore.ErrorDelay;
 			if (autoCorrupt && CPU_STEP_Count >= errorDelay)
 			{
 				CPU_STEP_Count = 0;
